Revert "Merge branch 'adding-react-router-architecture' into main"
This reverts commit d13bb0dc70ffaf0c25b1c1570fd24e3df17bc61f, reversing
changes made to 61616ed344473a63148c91aa7f0c35b77c2f234e.
