--- frontend/src/components/Buttons/_index.scss
@@ -1,2 +1,0 @@
-@forward "Buttons";
-@forward "CloseButton";
