+++ frontend/src/components/Navigation/HeaderNav.js
@@ -1,6 +1,6 @@
 // External Imports
 import PropTypes from "prop-types";
-import React, { Fragment, useState } from "react";
+import React, { useState, Fragment } from "react";
 
 // Internal Imports
 import * as utility from "../Utility/utils";
@@ -63,6 +63,7 @@ function HeaderNav({ logoDesktop, logoMobile, menu }) {
           </button>
         </div>
       </nav>
+
       <div aria-hidden="true"></div>
     </Fragment>
   );
