--- frontend/src/components/Navigation/_index.scss
@@ -1,2 +1,0 @@
-@forward "HeaderNav";
-@forward "InnerCopNav";
