--- frontend/src/components/components.js
@@ -1,19 +1,0 @@
-import { Button } from "./Buttons/Buttons";
-import { CloseButton } from "./Buttons/CloseButton";
-import { CopCard, InnerCopCard, InnerCopNavCard } from "./Cards/Cards";
-import { ScrollCarousel } from "./Carousel/ScrollCarousel";
-import { InnerCopNav } from "./Navigation/InnerCopNav";
-import { HeaderNav } from "./Navigation/HeaderNav";
-import { FooterNav } from "./Navigation/FooterNav";
-
-export {
-  Button,
-  CloseButton,
-  CopCard,
-  InnerCopCard,
-  InnerCopNavCard,
-  ScrollCarousel,
-  InnerCopNav,
-  HeaderNav,
-  FooterNav,
-};
