+++ frontend/src/index.js
@@ -1,2 +1,2 @@
-import App from "router/App";
+import App from "./components/App";
 import "./index.scss";
