--- jsconfig.json
@@ -1,6 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": "frontend/src"
-  },
-  "include": ["frontend/src"]
-}
